+++ frontend/src/Components/Home.js
@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react'
-
 import { SimpleGrid,Heading, Flex, Box, Button } from '@chakra-ui/react';
 import { useState,useRef} from 'react';
 import Startpage from './Startpage';
