+++ frontend/src/Components/Landing.css
@@ -12,7 +12,7 @@
 
 #signup{
     background-color: black;
-    height: 100vh;
+    height: 100%;
     width: 100vw;
     display: flex;
 }
@@ -23,8 +23,39 @@
 
     }
     #image{
+        height: 10px;
+        background-size: 100vw 10px;
+        background-color: white;   
+    }
+
+}
+
+
+@media screen and (max-width:800px) {
+
+    #image{
+        height: 25vh;
+        background-size: 100vw 25vh;
+        background-color: white;   
+    }
+
+}
+
+@media screen and (max-width:1100px) {
+
+    #image{
         height: 40vh;
         background-size: 100vw 40vh;
+        background-color: white;   
+    }
+
+}
+
+@media screen and (max-width:1500px) {
+
+    #image{
+        height: 50vh;
+        background-size: 100vw 50vh;
         background-color: white;   
     }
