+++ frontend/src/Components/Landingpage.js
@@ -1,15 +1,14 @@
-import { Box, Button, Flex} from '@chakra-ui/react'
-import React, { useState } from 'react'
-import { Link } from 'react-router-dom'
+import { Box,  Flex} from '@chakra-ui/react'
+import React from 'react'
 import './Landing.css'
-import Login from './Login'
-import Signup from './Signup'
+import Register from './Register'
+
 
 
 
 function Landingpage() {
 
-  const [login,setlogin] = useState(true);
+  
 
   return (
     <div id='main' >
@@ -26,53 +25,10 @@ function Landingpage() {
       </Box>
       </Flex>
       </div>
-      <div id='signup' >
-        <Flex color='white' display={{base:'none',md:'flex',lg:'flex'}} width={'50vw'} ml={'10rem'} mt={'20rem'} fontSize={'5rem'} >
-            Sign up to<br></br> Explore our <br></br> Website
-            
-        </Flex>
-        {
-          (login)?
-                <Box   as={'Flex'} borderRadius={'2rem'}  color={'black'} ml={{base:'3rem',md:'0rem',lg:'0rem'}} mr='10rem' backgroundColor={'whiteAlpha.700'} mt={'5rem'} mb='5rem' justifyContent={'center'} >
-                <Flex justifyContent={'center'} >
-                <Login/>
+      <Register/>
 
-                </Flex>
-                <Flex mb={'3rem'} justifyContent={'center'}  fontSize='1.3rem' >
-                      <Flex>Already a member?</Flex>
-                      <Link to={'/'} >
-                      <Flex onClick={()=>{setlogin(false)}}  >Log in</Flex>
-                      </Link>
-                      </Flex>  
-                </Box>:
 
-                  <Box  ml={{base:'3rem',md:'0rem',lg:'0rem'}} as={'Flex'} borderRadius={'2rem'}  color={'black'} mr='10rem' backgroundColor={'whiteAlpha.700'} mt={'5rem'} mb='5rem' justifyContent={'center'} >
-                  <Flex justifyContent={'center'} >
-                  <Signup/>
 
-                  </Flex>
-                  <Flex justifyContent={'center'}  fontSize='1.3rem' >
-                        <Flex>Not on Agroshield yet?</Flex>
-                        <Link to={'/'} >
-                        <Flex onClick={()=>{setlogin(true)}}  >Sign up</Flex>
-                        </Link>
-                        </Flex>  
-                  </Box>
-        }
-           
-      </div>
-
-        <div id='aboutus' >
-        <Flex justifyContent={'center'} color='black' mt={'10rem'} fontSize={'3rem'} >
-            Our Main motive is to provide proper solution to farmers
-          </Flex>
-          <Flex justifyContent={'center'} color='black' mt={'10rem'} fontSize={'3rem'} >
-            I am currently looking for internship contact me at kalvium.community 
-          </Flex>
-          <Flex justifyContent={'center'} color='black' fontSize={'5rem'} >
-            Made With Love ❤️ by anmol virk
-          </Flex>
-        </div>
 
 
     </div>
