+++ frontend/src/Components/navbar&footer.css
@@ -258,7 +258,7 @@ footer{
     justify-content: center;
     align-items: center;
     flex-direction: column;
-    
+    margin-bottom: -100px;
 }
 
 footer .social_icon, footer .menu{
